Imported activities and gitignored them

+++ commons/src/main/java/commons/Actions/Action.java
@@ -2,14 +2,11 @@ package commons.Actions;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-
-
-//CHECK IF LOMBOK IS ALLOWED
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.criteria.CriteriaBuilder;
 
 
 //tag to mark this object as a database object
@@ -20,7 +17,7 @@ public class Action {
 
     private final String title;
 
-    private final Integer consumption;
+    private final long consumption;
 
     private final String imagePath;
 
@@ -38,7 +35,7 @@ public class Action {
      * @param imagePath   the path to the image representing the image
      * @param source
      */
-    public Action(@JsonProperty("id") String id, @JsonProperty("image_path") String imagePath, @JsonProperty("title") String title, @JsonProperty("consumption_in_wh") Integer consumption,
+    public Action(@JsonProperty("id") String id, @JsonProperty("image_path") String imagePath, @JsonProperty("title") String title, @JsonProperty("consumption_in_wh") long consumption,
                   @JsonProperty("source") String source) {
 
         this.id = id;
@@ -47,8 +44,6 @@ public class Action {
         this.imagePath = imagePath;
         this.source = source;
     }
-
-
 
     /**
      * EMPTY CONSTRUCTOR BECAUSE FOR SOME GOD DAMNED REASON
@@ -88,7 +83,7 @@ public class Action {
      *
      * @return the consumption of the current action
      */
-    public Integer getConsumption() {
+    public long getConsumption() {
         return consumption;
     }
